+++ src/main/java/com/aoe/restapi/controller/auth/AuthRestControllerImpl.java
@@ -3,9 +3,9 @@ package com.aoe.restapi.controller.auth;
 import com.aoe.restapi.dto.LoginRequestDto;
 import com.aoe.restapi.model.entity.User;
 import com.aoe.restapi.model.service.user.UserService;
-import com.aoe.restapi.utility.Status.OperationStatus;
-import com.aoe.restapi.utility.Status.OperationStatusError;
-import com.aoe.restapi.utility.Status.OperationStatusSuccess;
+import com.aoe.restapi.utility.status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatusError;
+import com.aoe.restapi.utility.status.OperationStatusSuccess;
 import com.aoe.restapi.utility.auth.AuthUtil;
 import com.aoe.restapi.utility.auth.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
