+++ src/main/java/com/aoe/restapi/controller/user/UserRestControllerImpl.java
@@ -5,9 +5,9 @@ import com.aoe.restapi.model.entity.User;
 import com.aoe.restapi.model.service.base.BaseService;
 import com.aoe.restapi.model.service.user.UserService;
 import com.aoe.restapi.utility.http.ResponseUtil;
-import com.aoe.restapi.utility.Status.OperationStatus;
-import com.aoe.restapi.utility.Status.OperationStatusError;
-import com.aoe.restapi.utility.Status.OperationStatusSuccess;
+import com.aoe.restapi.utility.status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatusError;
+import com.aoe.restapi.utility.status.OperationStatusSuccess;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
