+++ src/main/java/com/aoe/restapi/model/service/project/ProjectServiceImpl.java
@@ -2,9 +2,9 @@ package com.aoe.restapi.model.service.project;
 
 import com.aoe.restapi.model.entity.Project;
 import com.aoe.restapi.model.service.base.crud.BaseCrudServiceImpl;
-import com.aoe.restapi.utility.Status.OperationStatus;
-import com.aoe.restapi.utility.Status.OperationStatusError;
-import com.aoe.restapi.utility.Status.OperationStatusSuccess;
+import com.aoe.restapi.utility.status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatusError;
+import com.aoe.restapi.utility.status.OperationStatusSuccess;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.http.HttpStatus;
