+++ src/main/java/com/aoe/restapi/model/service/relational/userdomain/UserDomainService.java
@@ -1,6 +1,6 @@
 package com.aoe.restapi.model.service.relational.userdomain;
 
-import com.aoe.restapi.utility.Status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatus;
 
 public interface UserDomainService {
     OperationStatus manageUserDomain(boolean link, int userId, int domainId);
