+++ src/main/java/com/aoe/restapi/model/service/relational/userissue/UserIssueService.java
@@ -1,6 +1,6 @@
 package com.aoe.restapi.model.service.relational.userissue;
 
-import com.aoe.restapi.utility.Status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatus;
 
 public interface UserIssueService {
     OperationStatus manageUserIssue(boolean link, int userId, int issueId);
