+++ src/main/java/com/aoe/restapi/model/service/relational/userissue/UserIssueServiceImpl.java
@@ -4,14 +4,13 @@ import com.aoe.restapi.model.entity.Issue;
 import com.aoe.restapi.model.entity.User;
 import com.aoe.restapi.model.service.Issue.IssueService;
 import com.aoe.restapi.model.service.user.UserService;
-import com.aoe.restapi.utility.Status.OperationStatus;
-import com.aoe.restapi.utility.Status.OperationStatusError;
-import com.aoe.restapi.utility.Status.OperationStatusSuccess;
+import com.aoe.restapi.utility.status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatusError;
+import com.aoe.restapi.utility.status.OperationStatusSuccess;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.HashSet;
 import java.util.Set;
 
 @Service
