+++ src/main/java/com/aoe/restapi/model/service/relational/userproject/UserProjectService.java
@@ -1,6 +1,6 @@
 package com.aoe.restapi.model.service.relational.userproject;
 
-import com.aoe.restapi.utility.Status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatus;
 
 public interface UserProjectService {
     OperationStatus manageUserInProject(boolean link, int userId, int projectId);
