+++ src/main/java/com/aoe/restapi/model/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.aoe.restapi.model.service.user;
 
 import com.aoe.restapi.model.service.base.crud.BaseCrudService;
-import com.aoe.restapi.utility.Status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatus;
 
 public interface UserService<T> extends BaseCrudService<T> {
     @Override
