+++ src/main/java/com/aoe/restapi/model/service/user/UserServiceImpl.java
@@ -3,9 +3,9 @@ package com.aoe.restapi.model.service.user;
 import com.aoe.restapi.model.dao.UserRepository;
 import com.aoe.restapi.model.entity.User;
 import com.aoe.restapi.model.service.base.crud.BaseCrudServiceImpl;
-import com.aoe.restapi.utility.Status.OperationStatus;
-import com.aoe.restapi.utility.Status.OperationStatusError;
-import com.aoe.restapi.utility.Status.OperationStatusSuccess;
+import com.aoe.restapi.utility.status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatusError;
+import com.aoe.restapi.utility.status.OperationStatusSuccess;
 import com.aoe.restapi.utility.auth.AuthUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
