+++ src/main/java/com/aoe/restapi/model/service/userrole/UserRoleService.java
@@ -1,7 +1,6 @@
 package com.aoe.restapi.model.service.userrole;
 
-import com.aoe.restapi.model.service.base.crud.BaseCrudService;
-import com.aoe.restapi.utility.Status.OperationStatus;
+import com.aoe.restapi.utility.status.OperationStatus;
 
 public interface UserRoleService<T> {
